pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["dataset*"]

[project]
name = "local-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "accelerate>=0.28.0",
    "datasets>=2.18.0",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",
    "numpy>=1.24.0",
    "wandb>=0.16.0",
    "bitsandbytes>=0.42.0",
]
